backend/requirements.txt
# FastAPI
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Environment
python-dotenv==1.0.1

# AI/ML
openai==1.57.0
exa-py==1.1.0
langgraph==0.2.56
langchain-openai==0.2.10
langchain-core==0.3.25

# Auth
pyjwt[crypto]==2.10.1
httpx==0.28.1

# Document Parsing
pypdf2==3.0.1
python-docx==1.1.2

# Database (optional, for backend DB access)
psycopg2-binary==2.9.10

